Make 15min analysis work. add a exposure dose viewer.

+++ wps_scripts/src/main/groovy/org/noise_planet/noisemodelling/wps/Matsim/Calculate_Matsim_Agent_Exposure.groovy
@@ -4,6 +4,14 @@ import geoserver.GeoServer
 import geoserver.catalog.Store
 import groovy.sql.Sql
 import org.geotools.jdbc.JDBCDataStore
+import org.jfree.chart.ChartFactory
+import org.jfree.chart.ChartPanel
+import org.jfree.chart.JFreeChart
+import org.jfree.chart.plot.XYPlot
+import org.jfree.data.category.CategoryDataset
+import org.jfree.data.general.DatasetUtils
+import org.jfree.data.xy.XYSeries
+import org.jfree.data.xy.XYSeriesCollection
 import org.locationtech.jts.geom.Coordinate
 import org.matsim.api.core.v01.Coord
 import org.matsim.api.core.v01.Id
@@ -30,6 +38,8 @@ import org.matsim.vehicles.Vehicle
 import org.noise_planet.noisemodelling.emission.EvaluateRoadSourceCnossos
 import org.noise_planet.noisemodelling.emission.RSParametersCnossos
 
+import javax.swing.JFrame
+import java.awt.BorderLayout
 import java.sql.Connection
 import java.sql.ResultSet
 import java.sql.Statement
@@ -127,6 +137,12 @@ def exec(Connection connection, input) {
         timeSlice = input["timeSlice"];
     }
 
+    agentId = 0;
+    if (input["plotOneAgentId"] as int != 0) {
+        agentId = input["plotOneAgentId"];
+    }
+    println agentId;
+
     String[] den = ["D", "E", "N"];
     String[] hourClock = ["0_1", "1_2", "2_3", "3_4", "4_5", "5_6", "6_7", "7_8", "8_9", "9_10", "10_11", "11_12", "12_13", "13_14", "14_15", "15_16", "16_17", "17_18", "18_19", "19_20", "20_21", "21_22", "22_23", "23_24"];
     String[] quarterClock = ["0h00_0h15", "0h15_0h30", "0h30_0h45", "0h45_1h00", "1h00_1h15", "1h15_1h30", "1h30_1h45", "1h45_2h00", "2h00_2h15", "2h15_2h30", "2h30_2h45", "2h45_3h00", "3h00_3h15", "3h15_3h30", "3h30_3h45", "3h45_4h00", "4h00_4h15", "4h15_4h30", "4h30_4h45", "4h45_5h00", "5h00_5h15", "5h15_5h30", "5h30_5h45", "5h45_6h00", "6h00_6h15", "6h15_6h30", "6h30_6h45", "6h45_7h00", "7h00_7h15", "7h15_7h30", "7h30_7h45", "7h45_8h00", "8h00_8h15", "8h15_8h30", "8h30_8h45", "8h45_9h00", "9h00_9h15", "9h15_9h30", "9h30_9h45", "9h45_10h00", "10h00_10h15", "10h15_10h30", "10h30_10h45", "10h45_11h00", "11h00_11h15", "11h15_11h30", "11h30_11h45", "11h45_12h00", "12h00_12h15", "12h15_12h30", "12h30_12h45", "12h45_13h00", "13h00_13h15", "13h15_13h30", "13h30_13h45", "13h45_14h00", "14h00_14h15", "14h15_14h30", "14h30_14h45", "14h45_15h00", "15h00_15h15", "15h15_15h30", "15h30_15h45", "15h45_16h00", "16h00_16h15", "16h15_16h30", "16h30_16h45", "16h45_17h00", "17h00_17h15", "17h15_17h30", "17h30_17h45", "17h45_18h00", "18h00_18h15", "18h15_18h30", "18h30_18h45", "18h45_19h00", "19h00_19h15", "19h15_19h30", "19h30_19h45", "19h45_20h00", "20h00_20h15", "20h15_20h30", "20h30_20h45", "20h45_21h00", "21h00_21h15", "21h15_21h30", "21h30_21h45", "21h45_22h00", "22h00_22h15", "22h15_22h30", "22h30_22h45", "22h45_23h00", "23h00_23h15", "23h15_23h30", "23h30_23h45", "23h45_24h00"];
@@ -146,22 +162,32 @@ def exec(Connection connection, input) {
 
     Sql sql = new Sql(connection)
 
-    sql.execute("DROP TABLE IF EXISTS " + outTableName)
-    sql.execute("CREATE TABLE " + outTableName + ''' ( 
-        ID integer PRIMARY KEY AUTO_INCREMENT, 
-        PERSON_ID varchar(255),
-        HOME_FACILITY_ID varchar(255),
-        HOME_GEOM geometry,
-        WORK_FACILITY_ID varchar(255),
-        LAEQ real
-    );''')
-
+    if (agentId == 0) {
+        sql.execute("DROP TABLE IF EXISTS " + outTableName)
+        sql.execute("CREATE TABLE " + outTableName + ''' ( 
+            ID integer PRIMARY KEY AUTO_INCREMENT, 
+            PERSON_ID varchar(255),
+            HOME_FACILITY_ID varchar(255),
+            HOME_GEOM geometry,
+            WORK_FACILITY_ID varchar(255),
+            LAEQ real
+        );''')
+    }
     Statement stmt = connection.createStatement();
 
+    int print = 1
+    int counter = 0
     for (Map.Entry<Id<Person>, Person> entry : persons.entrySet()) {
         String personId = entry.getKey().toString();
         Person person = entry.getValue();
         Plan plan = person.getSelectedPlan();
+
+        if (agentId > 0 && personId != agentId.toString()) {
+            continue;
+        }
+
+        XYSeries laeqSeries = new XYSeries("LAeq");
+        XYSeries doseSeries = new XYSeries("Dose");
 
         String homeId = "";
         String homeGeom = "";
@@ -239,7 +265,9 @@ def exec(Connection connection, input) {
                     ResultSet result = stmt.executeQuery(query);
                     while(result.next()) {
                         LAeq = 10 * Math.log10(Math.pow(10, LAeq / 10) + timeWeight * Math.pow(10, result.getDouble("LEQA") / 10));
+                        laeqSeries.add(h, result.getDouble("LEQA"));
                     }
+                    doseSeries.add(h, LAeq);
                 }
             }
         }
@@ -289,23 +317,48 @@ def exec(Connection connection, input) {
                             timeWeight = ((activityEnd - activityStart) / 900) / (24 * 4);
                         }
                     }
-                    String timeString = hourClock[h];
+                    String timeString = quarterClock[q];
                     String query = '''
                                     SELECT D.LEQA
                                     FROM ''' + dataTablePrefix + timeString + ''' D
                                     INNER JOIN RECEIVERS R
                                     ON D.IDRECEIVER = R.PK
-                                    WHERE FACILITY_ID = ''' + activityId + '''
+                                    WHERE FACILITY_ID = \'''' + activityId + '''\'
                                 '''
                     ResultSet result = stmt.executeQuery(query);
                     while(result.next()) {
                         LAeq = 10 * Math.log10(Math.pow(10, LAeq / 10) + timeWeight * Math.pow(10, result.getDouble("LEQA") / 10));
+                        laeqSeries.add(q, result.getDouble("LEQA"));
                     }
+                    doseSeries.add(q, LAeq);
                 }
             }
         }
-        println "Person id : " + personId + " : " + String.format("%.1f", LAeq) + " dB(A)"
 
-        sql.execute("INSERT INTO " + outTableName + " VALUES(NULL, \'" + personId + "\', \'" + homeId + "\', ST_GeomFromText(\'" + homeGeom + "\', 2154), \'" + workId + "\', " + LAeq + ")")
+        if (agentId > 0) {
+            XYSeriesCollection dataset = new XYSeriesCollection();
+            dataset.addSeries(laeqSeries);
+            dataset.addSeries(doseSeries);
+            JFreeChart chart = ChartFactory.createXYLineChart("AgentId : " + agentId, "Time", "Noise Level", dataset);
+            ChartPanel chartPanel = new ChartPanel(chart);
+
+            JFrame f = new JFrame("RESULT");
+            f.setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
+            f.setLayout(new BorderLayout(0, 5));
+            f.add(chartPanel, BorderLayout.CENTER);
+            f.pack();
+            f.setLocationRelativeTo(null);
+            f.setVisible(true);
+        }
+        if (counter == print) {
+            println "INFO Person # " + counter
+            print *= 4
+        }
+        // println "Person id : " + personId + " : " + String.format("%.1f", LAeq) + " dB(A)"
+
+        if (agentId == 0) {
+            sql.execute("INSERT INTO " + outTableName + " VALUES(NULL, \'" + personId + "\', \'" + homeId + "\', ST_GeomFromText(\'" + homeGeom + "\', 2154), \'" + workId + "\', " + LAeq + ")")
+        }
+        counter++;
     }
 }
